+++ src/main/java/org/patryk3211/powergrid/electricity/solarpanel/SolarPanelBearingContraption.java
@@ -19,7 +19,6 @@ public class SolarPanelBearingContraption extends BearingContraption {
     protected int panelBlocks;
     protected Vector3d panelNormal;
 
-
     public SolarPanelBearingContraption() {
         super();
     }
